requirements.txt
transformers
flash-attn
accelerate
bitsandbytes
huggingface_hub
openai
flask
asyncio
gradio
python-dotenv
